requirements.txt
autopep8==1.5.7
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.7.0
python-dotenv==0.19.1
rsa==4.7.2
Telethon==1.23.0
toml==0.10.2
